refactor: rename component to StudentDetailsByID and make it independent to display student details by ID
