--- src/Enrollment/Student/Pages/StudentEmail/StudentEmailWrapper.test.tsx
@@ -1,39 +1,0 @@
-import { render, screen } from "@testing-library/react";
-import { MemoryRouter, Route, Routes } from "react-router-dom";
-import { Student } from "../../studentData/studentTypes";
-import { Email } from "../../studentData/email";
-import StudentEmailWrapper from "./StudentEmailWrapper";
-
-const testStudents: Student[] = [
-  {
-    id: "s1",
-    name: "Alice",
-    email: new Email("alice@example.com"),
-    group: "G1",
-    role: "Student",
-    section: "A",
-    imageUrl: "https://placehold.co/100",
-    notes: "Top student",
-    isLoopEnrolled: true,
-    isGithubEnrolled: false,
-  },
-];
-
-describe("StudentEmailWrapper", () => {
-  it("renders StudentEmail when student is found", () => {
-    render(
-      <MemoryRouter initialEntries={["/students/email/s1"]}>
-        <Routes>
-          <Route
-            path="/students/email/:id"
-            element={
-              <StudentEmailWrapper studentGetter={() => testStudents} />
-            }
-          />
-        </Routes>
-      </MemoryRouter>
-    );
-    expect(screen.getByText("Alice")).toBeInTheDocument();
-    expect(screen.getByText("alice@example.com")).toBeInTheDocument();
-  });
-});
