+++ src/Enrollment/Student/studentData/findStudentById.test.ts
@@ -1,5 +1,5 @@
 import { findStudentById } from "./findStudentById";
-import { Email } from "./email";
+
 import { Student } from "./studentTypes";
